refactor: component for equipment filters itself
